Add a user for the mediacenter
I wanted to create a user for the mediacenter, that has only the stuff
used by the mediacenter (and e.g. no other dotfiles) installed.

The user will be automatically logged in by `gdm` (I could have used
`nodm`, but `gdm` was already installed on the system) and therefore
needs no password to be set. This has the benefit, that this is a
"local only" user, i.e. that one cannot login to it via `ssh`.