Install `kodi`
I will use `kodi` as the mediacenter of choice, since it has the most
support. At the current time `plasma-bigpicture` seems to be not there
yet.

I also made it the default session for `mediacenter`, as I want it to
start automatically. I chose `copy` over `lineinfile`, since the file
is only created at the first login of the user and would therefore fail
on fresh installs.