$ vcsh mediacenter ls-files | xargs npx prettier -w
My `nvim` config autoformats everything that `prettier` supports with
`prettier` now. This was a bit annoying, since the formatting did not
match mine, so I had to make sure to `:noa w` all the time.

Since I love autoformatters I just reformatted this repo with the
command above and can now leverage `prettier` in the future :)